fix: update tests for createSlice refactor
