+++ src/containers/CriterionContainer/CriterionFeedback.test.jsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 
-import actions from 'data/actions';
-import selectors from 'data/selectors';
+import { actions, selectors } from 'data/redux';
 import {
   feedbackRequirement,
   gradeStatuses,
